Initial this project with a basic game.
I initial this project with a basic game that allow user to act as a
spirit in the browser, we can move our character by pressing down the "WASD"
key for now.
Through this example, I have demonstrated how the communication is done between
clients and servers, and we will base on that to improve this
game.